add loader and lombok

+++ src/main/java/gh7/web/service/ImmigrantRepository.java
@@ -3,5 +3,5 @@ package gh7.web.service;
 import gh7.domain.Immigrant;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ImmigrantRepository extends CrudRepository<Immigrant, Integer> {
+public interface ImmigrantRepository extends CrudRepository<Immigrant, Long> {
 }
